refactor(prisma): wrap the Prisma service within a Module
