+++ backend/src/app.module.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from './prisma.service';
+import { PrismaService } from './prisma/prisma.service';
 import { Module } from '@nestjs/common';
 import { ThrottlerModule } from '@nestjs/throttler';
 import { AppController } from './app.controller';
