+++ backend/src/auth/auth.module.ts
@@ -8,11 +8,11 @@ import { LocalStrategy } from './strategies/local.strategy';
 import { NovuModule } from '../novu/novu.module';
 import { VerifyStrategy } from './strategies/verify.strategy';
 import { AppConfigService } from '../config/config.service';
-import { UsersModule } from '../users/users.module';
+import { PrismaModule } from '../prisma/prisma.module';
 
 @Module({
   imports: [
-    UsersModule,
+    PrismaModule,
     PassportModule,
     JwtModule.registerAsync({
       inject: [AppConfigService],
