+++ backend/src/prisma/prisma.service.spec.ts
@@ -1,0 +1,18 @@
+import { PrismaService } from './prisma.service';
+import { Test } from '@nestjs/testing';
+
+describe('PrismaService', () => {
+  let service: PrismaService;
+
+  beforeEach(async () => {
+    const app = await Test.createTestingModule({
+      providers: [PrismaService],
+    }).compile();
+
+    service = app.get<PrismaService>(PrismaService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
