+++ backend/src/shared/decorators/unique.decorator.ts
@@ -6,7 +6,7 @@ import {
   ValidationArguments,
 } from 'class-validator';
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../../prisma.service';
+import { PrismaService } from '../../prisma/prisma.service';
 
 @ValidatorConstraint({ name: 'Unique', async: true })
 @Injectable()
